+++ EA/src/Network.h
@@ -18,7 +18,7 @@ public:
    Network(Network *network);
 
    // Constructor used for testing
-   Network(int **matrix, int *weights, int size);
+   Network(double **matrix, double *weights, int size);
 
    virtual ~Network();
 
@@ -27,17 +27,17 @@ public:
 
    inline int size() const { return m_size; }
 
-   inline int node_weight(int node) const { return m_node_weights[node]; }
+   inline double node_weight(int node) const { return m_node_weights[node]; }
 
-   inline int link_weight(int row, int column) const { return m_link_weights[row][column]; }
+   inline double link_weight(int row, int column) const { return m_link_weights[row][column]; }
 
    /* 
     * Determines whether the network has a path from node1 to node2
     * such that every edge along the path is no less than the given weight
     */
-   bool path_exists(int node1, int node2, int weight);
+   bool path_exists(int node1, int node2, double weight);
 
-   bool embed_path(int node1, int node2, int weight);
+   bool embed_path(int node1, int node2, double weight);
 
 private:
    /*
@@ -52,7 +52,7 @@ private:
     *         by 'weight'
     *
     */
-   void reduce_capacity(int node1, int node2, int *node_ancestor, int weight);
+   void reduce_capacity(int node1, int node2, int *node_ancestor, double weight);
 
    /*  
     * Initializes m_size based on the number of nodes defined in the input_file
@@ -65,9 +65,9 @@ private:
     */
    void init_size(FILE *input_file);
 
-   int **m_link_weights; // graph representation, with edge weights
-   int  *m_node_weights;     // weights on graph nodes
-   int   m_size;        // number of graph nodes
+   double **m_link_weights; // graph representation, with edge weights
+   double  *m_node_weights;     // weights on graph nodes
+   int      m_size;        // number of graph nodes
 };
 #endif
